Changes done

+++ Assignments/22Assg.c
@@ -3,25 +3,25 @@
 int main()
 {
     char op;
-    double first, second;
+    float first, second;
     printf("Enter an operator (+, -, *, /): ");
     scanf("%c", &op);
     printf("Enter two operands: ");
-    scanf("%lf %lf", &first, &second);
+    scanf("%f %f", &first, &second);
 
     switch (op)
     {
     case '+':
-        printf("%.1lf + %.1lf = %.1lf", first, second, first + second);
+        printf("%.1f + %.1f = %.1f", first, second, first + second);
         break;
     case '-':
-        printf("%.1lf - %.1lf = %.1lf", first, second, first - second);
+        printf("%.1f - %.1f = %.1f", first, second, first - second);
         break;
     case '*':
-        printf("%.1lf * %.1lf = %.1lf", first, second, first * second);
+        printf("%.1f * %.1f = %.1f", first, second, first * second);
         break;
     case '/':
-        printf("%.1lf / %.1lf = %.1lf", first, second, first / second);
+        printf("%.1f / %.1f = %.1f", first, second, first / second);
         break;
     // operator doesn't match any case constant
     default:
